Fix compilation issues

+++ PyBoltz/MBSorts.pyx
@@ -6,13 +6,13 @@ import  numpy as np
 
 cdef long long MBSort(int I, double R2, int IE,Boltz Object):
     """
-    This function selects collision type from collision array by binary step sampling reduces sampling range to within 4 
+    This function selects collision type from collision array by binary step sampling reduces sampling range to within 4
     positions in array output =  i ( position within 4 of correct value).
-    :param I: 
-    :param R2: 
-    :param IE: 
-    :param Object: 
-    :return: a new index 
+    :param I:
+    :param R2:
+    :param IE:
+    :param Object:
+    :return: a new index
     """
     cdef long long iStep,Increment
     cdef int K
@@ -30,7 +30,7 @@ cdef long long MBSort(int I, double R2, int IE,Boltz Object):
         if I <= Object.NumMomCrossSectionPointsNT:
             if Object.CollisionFrequencyNT[IE][I - 1] < R2:
                 Increment = Increment + iStep
-        iStep = iStep / 2
+        iStep = long(iStep / 2)
     if I==0:
         return I
     return I - 1
@@ -38,13 +38,13 @@ cdef long long MBSort(int I, double R2, int IE,Boltz Object):
 
 cdef long long MBSortT(int GasIndex, int I, double R2, int IE,Boltz Object):
     """
-    This function selects collision type from collision array by binary step sampling reduces sampling range to within 4 
+    This function selects collision type from collision array by binary step sampling reduces sampling range to within 4
     positions in array output =  i ( position within 4 of correct value).
-    :param I: 
-    :param R2: 
-    :param IE: 
-    :param Object: 
-    :return: a new index 
+    :param I:
+    :param R2:
+    :param IE:
+    :param Object:
+    :return: a new index
     """
 
     cdef long long iStep,Increment
@@ -62,7 +62,7 @@ cdef long long MBSortT(int GasIndex, int I, double R2, int IE,Boltz Object):
         if I <= Object.NumMomCrossSectionPoints[GasIndex]-1:
             if Object.CollisionFrequency[GasIndex][IE][I-1] < R2:
                 Increment = Increment + iStep
-        iStep = iStep / 2
+        iStep = long(iStep / 2)
     if I==0:
         return I
     return I -1
